refactor: Fix folder structures
