+++ src/features/discovery-detail/ui/vote-dialog.tsx
